Create ECHO.java
If you have ever visited a canyon or a cave, you may have yelled and heard the echo of your own voice. In this problem, you should simulate that effect:

Hello! Hello! Hello!
Your program will be given as input a single word, and it should print out that word three times, separating the words with spaces.